RootClickAction renamed to ClickRootAction
